Removed unused data properties. Moved api result status, network usage functions to separate package.
